fix: rename decrypt-udf to cipher-udf
