+++ src/service/search/datafusion/udf/mod.rs
@@ -22,9 +22,9 @@ pub(crate) mod arrjoin_udf;
 pub(crate) mod arrsort_udf;
 pub(crate) mod arrzip_udf;
 pub(crate) mod cast_to_arr_udf;
+pub(crate) mod cipher_udf;
 pub(crate) mod date_format_udf;
 pub(crate) mod fuzzy_match_udf;
-pub(crate) mod decrypt_udf;
 pub(crate) mod histogram_udf;
 pub(crate) mod match_all_udf;
 pub(crate) mod regexp_matches_udf;
